requirements.txt
python-decouple
telethon
schedule
bs4
requests
telethon
selenium